bugfix in neighsearch errhandler, bugfix in dhdt routine
git-svn-id: https://isis2.unibe.ch/repos/sphlatch@163 4efe6d44-883f-0410-9ed0-acde54f1c0bb

+++ src/rankspace.h
@@ -324,7 +324,7 @@ void findNeighbours(const size_t _curPartIndex,
                     {
                       noNeighbours++;
 #ifdef SPHLATCH_CHECKNONEIGHBOURS
-                      if ( noNeighbours > noMaxNeighs )
+                      if ( static_cast<int>(noNeighbours) > noMaxNeighs )
                         throw TooManyNeighs( _curPartIndex,
                                              noNeighbours,
                                              neighbourList,
